refactor: remove unused code to improve coverage
