+++ src/main/java/artgallery/hsboxoffice/configuration/ApiError.java
@@ -1,0 +1,28 @@
+package artgallery.hsboxoffice.configuration;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class ApiError {
+
+    private HttpStatus status;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy HH:mm:ss")
+    private LocalDateTime timestamp;
+    private String message;
+    private String debugMessage;
+
+    private ApiError() {
+        timestamp = LocalDateTime.now();
+    }
+
+    public ApiError(HttpStatus status, Throwable ex) {
+        this();
+        this.status = status;
+        this.message = "Unexpected error";
+        this.debugMessage = ex.getMessage();
+    }
+}
