--- src/main/java/artgallery/hsboxoffice/controller/ApiError.java
@@ -1,71 +1,0 @@
-package artgallery.hsboxoffice.controller;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-public class ApiError {
-
-    private HttpStatus status;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy HH:mm:ss")
-    private LocalDateTime timestamp;
-    private String message;
-    private String debugMessage;
-    private List<ApiSubError> subErrors;
-
-    private ApiError() {
-        timestamp = LocalDateTime.now();
-    }
-
-    public ApiError(HttpStatus status) {
-        this();
-        this.status = status;
-    }
-
-    public ApiError(HttpStatus status, Throwable ex) {
-        this();
-        this.status = status;
-        this.message = "Unexpected error";
-        this.debugMessage = ex.getMessage();
-    }
-
-    public ApiError(HttpStatus status, String message) {
-        this();
-        this.status = status;
-        this.message = message;
-        this.debugMessage = null;
-    }
-
-    public ApiError(HttpStatus status, String message, Throwable ex) {
-        this();
-        this.status = status;
-        this.message = message;
-        this.debugMessage = ex.getMessage();
-    }
-}
-
-abstract class ApiSubError {
-
-}
-
-@Data
-@EqualsAndHashCode(callSuper = false)
-@AllArgsConstructor
-class ApiValidationError extends ApiSubError {
-    private String object;
-    private String field;
-    private Object rejectedValue;
-    private String message;
-
-    ApiValidationError(String object, String message) {
-        this.object = object;
-        this.message = message;
-    }
-}
