--- src/main/java/artgallery/hsboxoffice/exception/DatabaseConflictException.java
@@ -1,7 +1,0 @@
-package artgallery.hsboxoffice.exception;
-
-public class DatabaseConflictException extends Exception {
-    public DatabaseConflictException(String message) {
-        super(message);
-    }
-}
